Added Getopt::Long and some basic runtime arguments.
Added html output support
Added PDF output support, this though we may want to switch to PDF::API2 rather than PDF::Create as it appears that PM supports more of the capabilities that we are looking for.  PDF output is not pretty and does not yet contain any graph data.